requirements.txt
# Core Dependencies
python-frontmatter==1.0.1
pypandoc-binary==1.13
markdown==3.6
weasyprint==62.3
mistune==3.0.2

# Syntax Highlighting
pygments==2.18.0

# Web Framework
flask==3.0.3
gunicorn==22.0.0
werkzeug==3.0.3

# OAuth2 Authentication
Flask-Dance>=7.1.0

# Google API Integration
google-api-python-client>=2.149.0
google-auth>=2.35.0

# Markdown to Google Docs Conversion
markgdoc>=1.0.1

# Utilities
python-dotenv==1.0.0
python-docx==1.1.2

# HTML Conversion
bleach>=6.3.0
beautifulsoup4>=4.14.2
requests>=2.32.0
